Social: Add a 'more info' link to the editor upgrade nudge. (#27617)
* Social: Add a 'more info' link to the editor upgrade nudge.

The upgrade link in the editor drops the customer directly in the
checkout screen. This PR adds a more info link that takes them to the
main jetpack.com pricing page, with the Jetpack Social plan details open
in a modal.

If they checkout from here they will be redirected back to the editor.

* Update the more info link to open in a pop-up.

* Update the more info link to open in a new tab

This also updates the redirect, so the user is sent back to a plugin
admin page after they have purcchased a plan.

* changelog

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3648016121