Social: Prevent the metadata updating on autosave (#29263)
* Social: Prevent the metadata updating on autosave

A problem was reported where connections were being incorrectly skipped.
It appears that this happens when there is a delay in sync and the post
is autosaved in the meantime. If that happens we currently update the
connection metadata based on the connections being disabled (which
happens immediately after publish).

This is a quick fix for the problem which prevents us from updating the
metadata if we're performing an autosave. This is the same behaviour as
in the classic editor.

* Update package versions

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4424340065