Publicize: Reinstate is_social_image_generator_enabled to avoid fatals (#29952)
* Publicize: Reinstate is_social_image_generator_enabled to avoid fatals

When a previous version of Jetpack is active with a newer version of
Social, a fatal error would occur when loading the editor.

This is because we removed the is_social_image_generator_enabled method
in #29529 and should have left it for backwards compatibility. This
change reinstates it as an alias for has_social_image_generator_feature.

* Fix package version.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4625575668