Social: Swap to using the My Jetpack Endpoint for product information (#42812)
* Social: Always register the product info endpoint.

With only Jetpack active, the admin page attempts to use the
social-product-info endpoint to retrieve the planning pricing, but it's
not regsitered if the Publicize module isn't active.

This would also be a problem with the new version of the Social plugin,
but the current one registers the endpoints regardless.

This change ensures that we have the endpoint registered, even if the
Publicize module is inactive.

* Change to only register the endpoint on non-WordPress.com sites.

* Change to using the My Jetpack endpoint to get the pricing information

* Deprecate the social-product-info endpoint.

* Fix the useProductInfo test

* Restore the order of the endpoint registration.

* Add deprecated annotation

* Test the deprecation notice for the endpoint

* Switch to using a string to capture the deprecation notice.

---------

Co-authored-by: Gergely Juhasz <gergely.juhasz@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/14375625010

Upstream-Ref: Automattic/jetpack@14a40a3b8f75f21ea5a3bcaca08e7ee9680c2e91